do the same for tests
